+++ base/range.jl
@@ -728,27 +728,6 @@ end
 -(r::StepRangeLen) = StepRangeLen(-r.ref, -r.step, length(r), r.offset)
 -(r::LinSpace) = LinSpace(-r.start, -r.stop, length(r))
 
-+(x::Real, r::AbstractUnitRange) = range(x + first(r), length(r))
-# For #18336 we need to prevent promotion of the step type:
-+(x::Number, r::AbstractUnitRange) = range(x + first(r), step(r), length(r))
-+(x::Number, r::Range) = (x+first(r)):step(r):(x+last(r))
-function +(x::Number, r::StepRangeLen)
-    newref = x + r.ref
-    StepRangeLen{eltype(newref),typeof(newref),typeof(r.step)}(newref, r.step, length(r), r.offset)
-end
-function +(x::Number, r::LinSpace)
-    LinSpace(x + r.start, x + r.stop, r.len)
-end
-+(r::Range, x::Number) = x + r  # assumes addition is commutative
-
--(x::Number, r::Range)      = (x-first(r)):-step(r):(x-last(r))
--(x::Number, r::StepRangeLen) = +(x, -r)
-function -(x::Number, r::LinSpace)
-    LinSpace(x - r.start, x - r.stop, r.len)
-end
-
--(r::Range, x::Number) = +(-x, r)
-
 *(x::Number, r::Range)        = range(x*first(r), x*step(r), length(r))
 *(x::Number, r::StepRangeLen) = StepRangeLen(x*r.ref, x*r.step, length(r), r.offset)
 *(x::Number, r::LinSpace)     = LinSpace(x * r.start, x * r.stop, r.len)
@@ -760,8 +739,53 @@ end
 /(r::Range, x::Number)        = range(first(r)/x, step(r)/x, length(r))
 /(r::StepRangeLen, x::Number) = StepRangeLen(r.ref/x, r.step/x, length(r), r.offset)
 /(r::LinSpace, x::Number)     = LinSpace(r.start / x, r.stop / x, r.len)
+# also, separate in case of noncommutative multiplication (division)
+\(x::Number, r::Range)        = range(x\first(r), x\step(r), x\length(r))
+\(x::Number, r::StepRangeLen) = StepRangeLen(x\r.ref, x\r.step, length(r), r.offset)
+\(x::Number, r::LinSpace)     = LinSpace(x \ r.start, x \ r.stop, r.len)
 
-/(x::Number, r::Range) = [ x/y for y=r ]
+## scalar-range broadcast operations ##
+
+broadcast(::typeof(-), r::OrdinalRange) = range(-first(r), -step(r), length(r))
+broadcast(::typeof(-), r::StepRangeLen) = StepRangeLen(-r.ref, -r.step, length(r), r.offset)
+broadcast(::typeof(-), r::LinSpace) = LinSpace(-r.start, -r.stop, length(r))
+
+broadcast(::typeof(+), x::Real, r::AbstractUnitRange) = range(x + first(r), length(r))
+# For #18336 we need to prevent promotion of the step type:
+broadcast(::typeof(+), x::Number, r::AbstractUnitRange) = range(x + first(r), step(r), length(r))
+broadcast(::typeof(+), x::Number, r::Range) = (x+first(r)):step(r):(x+last(r))
+function broadcast(::typeof(+), x::Number, r::StepRangeLen)
+    newref = x + r.ref
+    StepRangeLen{eltype(newref),typeof(newref),typeof(r.step)}(newref, r.step, length(r), r.offset)
+end
+function broadcast(::typeof(+), x::Number, r::LinSpace)
+    LinSpace(x + r.start, x + r.stop, r.len)
+end
+broadcast(::typeof(+), r::Range, x::Number) = broadcast(+, x, r)  # assumes addition is commutative
+
+broadcast(::typeof(-), x::Number, r::Range)      = (x-first(r)):-step(r):(x-last(r))
+broadcast(::typeof(-), x::Number, r::StepRangeLen) = broadcast(+, x, -r)
+function broadcast(::typeof(-), x::Number, r::LinSpace)
+    LinSpace(x - r.start, x - r.stop, r.len)
+end
+
+broadcast(::typeof(-), r::Range, x::Number) = broadcast(+, -x, r)  # assumes addition is commutative
+
+broadcast(::typeof(*), x::Number, r::Range)        = range(x*first(r), x*step(r), length(r))
+broadcast(::typeof(*), x::Number, r::StepRangeLen) = StepRangeLen(x*r.ref, x*r.step, length(r), r.offset)
+broadcast(::typeof(*), x::Number, r::LinSpace)     = LinSpace(x * r.start, x * r.stop, r.len)
+# separate in case of noncommutative multiplication
+broadcast(::typeof(*), r::Range, x::Number)        = range(first(r)*x, step(r)*x, length(r))
+broadcast(::typeof(*), r::StepRangeLen, x::Number) = StepRangeLen(r.ref*x, r.step*x, length(r), r.offset)
+broadcast(::typeof(*), r::LinSpace, x::Number)     = LinSpace(r.start * x, r.stop * x, r.len)
+
+broadcast(::typeof(/), r::Range, x::Number)        = range(first(r)/x, step(r)/x, length(r))
+broadcast(::typeof(/), r::StepRangeLen, x::Number) = StepRangeLen(r.ref/x, r.step/x, length(r), r.offset)
+broadcast(::typeof(/), r::LinSpace, x::Number)     = LinSpace(r.start / x, r.stop / x, r.len)
+# also, separate in case of noncommutative multiplication (division)
+broadcast(::typeof(\), x::Number, r::Range)        = range(x\first(r), x\step(r), x\length(r))
+broadcast(::typeof(\), x::Number, r::StepRangeLen) = StepRangeLen(x\r.ref, x\r.step, length(r), r.offset)
+broadcast(::typeof(\), x::Number, r::LinSpace)     = LinSpace(x \ r.start, x \ r.stop, r.len)
 
 # promote eltype if at least one container wouldn't change, otherwise join container types.
 el_same(::Type{T}, a::Type{<:AbstractArray{T,n}}, b::Type{<:AbstractArray{T,n}}) where {T,n}   = a
@@ -945,3 +969,6 @@ function +(r1::StepRangeLen{T,S}, r2::StepRangeLen{T,S}) where {T,S}
 end
 
 -(r1::StepRangeLen, r2::StepRangeLen) = +(r1, -r2)
+
+broadcast(::typeof(+), r1::Range, r2::Range) = r1 + r2
+broadcast(::typeof(-), r1::Range, r2::Range) = r1 - r2